add conventions; use CODATA G
