+++ gwspace/constants.py
@@ -19,7 +19,7 @@ INVSQRT2 = 0.707106781186547524400844362104849039
 INVSQRT3 = 0.577350269189625764509148780501957455
 INVSQRT6 = 0.408248290463863016366214012450981898
 C_SI = 299792458.
-G_SI = 6.67259e-11
+G_SI = 6.67430e-11
 MSUN_SI = 1.988546954961461467461011951140572744e30
 MTSUN_SI = 4.9254923218988636432342917247829673e-6
 PC_SI = 3.085677581491367278913937957796471611e16
